+++ includes/goal/edit-goal.php
@@ -3,6 +3,11 @@
 // This script is always a request.
 include_once("../constants.php");
 include_once(CLASS_PATH . "Goal.php");
+include_once(CLASS_PATH . "Category.php");
+
+if (!isset($_SESSION)) {
+    session_start();
+}
 
 $goalId = $_GET["goalId"];
 
@@ -14,9 +19,21 @@ if (isset($_GET["name"])) {
 $categoryId = null;
 if (isset($_GET["categoryId"])) {
     $categoryId = $_GET["categoryId"];
+    $category = new Category($categoryId);
+
+    if (!$category->verifyCategoryOwnership($_SESSION["user_id"])) {
+        echo "unverified";
+        exit();
+    }
 }
 
 $goal = new Goal($goalId);
+
+if (!$goal->verifyGoalOwnership($_SESSION["user_id"])) {
+    echo "unverified";
+    exit();
+}
+
 $goal->editGoal($name, $categoryId);
 
 // add support for editing steps.
@@ -29,6 +46,11 @@ if (isset($_GET["steps"])) {
         if (count($step) == 3) {
             $name = $step[0];
             $date = $step[1];
+
+            if (!$goal->verifyStepOwnership($step[2])) {
+                echo "unverified";
+                exit();
+            }
 
             if ($name == null && $date == null) {
                 $goal->deleteStep($step[2]);
